Add error messaging for error when command is incorrect, and tests.
This creates a new error type for flag usage, such that we can tell if 
it was bad input or an actual bot failure. In the case of an actual bot 
failure we log it to console, but in the case of user error we reply 
with a specific error message. This also adds some testing on the 
replied message.